Address review round 2: version dirs, python submodule split, merge main
- Rename pickle fixture dir 1.X/ -> 1.3/ (the unreleased breaking-changes
  version) and add a 1.2/ dir with the same fixtures regenerated under
  discopy 1.2.2 (the latest release).
- test_pickle_version_compatibility now checks the 1.3 reference against
  both older versions (0.6 and 1.2).
- Split test/python.py into one file per submodule under test/python/
  (multiplicative, finset, additive), matching the grammar/quantum/drawing
  convention; note python in AGENTS.md's test-layout description.

Also merged origin/main to pick up the Functor ob/ar -> ob_map/ar_map
rename (#369).
