+++ test/fixtures/pickles/1.3/quantum.zx.Diagram.pickle
