+++ test/fixtures/pickles/1.3/rigid.Box.pickle
